Disount Acceptance Tests, curtesy of my french handsome friend, all pass

+++ src/main/java/Application/DTOs/DiscountDTO.java
@@ -181,7 +181,7 @@ public class DiscountDTO {
         return discountPercentage;
     }
 
-    public void setDiscountPercentage(Float discountPercentage) {
+    public void setDiscountPercentage(double discountPercentage) {
         this.discountPercentage = discountPercentage;
     }
